+++ api/src/main/java/au/gov/digitalhealth/lingo/service/ProductCalculationService.java
@@ -312,7 +312,8 @@ public abstract class ProductCalculationService<T extends ProductDetails> {
             subordinateNode =
                 getNodeGeneratorService().lookUpNode(branch, linkedConcept, level, null).join();
             subordinateNode.setOriginalNode(
-                new OriginalNode(subordinateNode.cloneNode(), null, true));
+                new OriginalNode(
+                    subordinateNode.cloneNode(), null, true, modelConfiguration.getModuleId()));
             nodeCache.put(linkedConcept.getConceptId(), subordinateNode);
           }